Move helper methods out of rest util
