+++ codalab/rest/bundle.py
@@ -1,6 +1,7 @@
 import httplib
 import mimetypes
 import os
+import re
 import sys
 import time
 
@@ -13,9 +14,15 @@ from marshmallow import (
 )
 from marshmallow_jsonapi import Schema, fields
 
-from codalab.bundles import BUNDLE_SUBCLASSES, get_bundle_subclass
-from codalab.common import UsageError, State
-from codalab.lib import spec_util, zip_util, worksheet_util
+from codalab.bundles import BUNDLE_SUBCLASSES, get_bundle_subclass, PrivateBundle
+from codalab.common import State, UsageError
+from codalab.lib import (
+    bundle_util,
+    canonicalize,
+    spec_util,
+    zip_util,
+    worksheet_util,
+)
 from codalab.lib.server_util import (
     bottle_patch as patch,
     json_api_include,
@@ -25,6 +32,7 @@ from codalab.lib.server_util import (
     query_get_type,
 )
 from codalab.lib.spec_util import validate_uuid, validate_child_path
+from codalab.model.tables import GROUP_OBJECT_PERMISSION_READ
 from codalab.objects.permission import (
     check_bundles_have_all_permission,
     check_bundles_have_read_permission,
@@ -32,13 +40,10 @@ from codalab.objects.permission import (
     PermissionSpec,
 )
 from codalab.rest.users import UserSchema
+from codalab.rest.worksheets import get_worksheet_uuid
 from codalab.rest.util import (
-    delete_bundles as _delete_bundles,
-    get_bundle_infos,
+    local_bundle_client_compatible,
     get_resource_ids,
-    resolve_bundle_specs,
-    resolve_owner_in_keywords,
-    set_bundle_permissions as _set_bundle_permissions,
 )
 from codalab.server.authenticated_plugin import AuthenticatedPlugin
 
@@ -95,6 +100,7 @@ class BundleSchema(Schema):
     children = fields.Relationship(many=True, type_='bundles', id_field='uuid', include_data=True)
     group_permissions = fields.Relationship(many=True, type_='bundle-permissions', id_field='id', include_data=True)
     host_worksheets = fields.List(fields.Dict)
+    args = fields.String()
 
     # Bundle permission of the authenticated user for convenience, read-only
     permission = fields.Integer()
@@ -106,12 +112,13 @@ class BundleSchema(Schema):
 
 CREATE_RESTRICTED_FIELDS = ('id', 'uuid', 'data_hash', 'state', 'owner',
                             'children', 'group_permissions', 'host_worksheets',
-                            'permission', 'permission_spec')
+                            'args', 'permission', 'permission_spec')
 
 
 UPDATE_RESTRICTED_FIELDS = ('command', 'data_hash', 'state', 'dependencies',
                             'children', 'group_permissions', 'host_worksheets',
-                            'permission', 'permission_spec', 'bundle_type')
+                            'args', 'permission', 'permission_spec',
+                            'bundle_type')
 
 
 #############################################################
@@ -120,14 +127,14 @@ UPDATE_RESTRICTED_FIELDS = ('command', 'data_hash', 'state', 'dependencies',
 
 
 @get('/bundles/<uuid:re:%s>' % spec_util.UUID_STR, apply=AuthenticatedPlugin())
-def fetch_bundle(uuid):
-    document = fetch_bundles_helper([uuid])
-    document['data'] = document['data'][0]
+def _fetch_bundle(uuid):
+    document = build_bundles_document([uuid])
+    document['data'] = document['data'][0]  # Flatten data list
     return document
 
 
 @get('/bundles', apply=AuthenticatedPlugin())
-def fetch_bundles():
+def _fetch_bundles():
     """
     Fetch bundles by bundle specs OR search keywords.
     """
@@ -151,10 +158,10 @@ def fetch_bundles():
     if not isinstance(bundle_uuids, list):
         return json_api_meta({}, {'result': bundle_uuids})
 
-    return fetch_bundles_helper(bundle_uuids)
+    return build_bundles_document(bundle_uuids)
 
 
-def fetch_bundles_helper(bundle_uuids):
+def build_bundles_document(bundle_uuids):
     descendant_depth = query_get_type(int, 'list-descendants', None)
 
     bundles_dict = get_bundle_infos(
@@ -202,7 +209,7 @@ def fetch_bundles_helper(bundle_uuids):
 
 
 @post('/bundles', apply=AuthenticatedPlugin())
-def create_bundles():
+def _create_bundles():
     """
     Bulk create bundles.
     """
@@ -248,7 +255,7 @@ def create_bundles():
         # Inherit worksheet permissions
         group_permissions = local.model.get_group_worksheet_permissions(
             request.user.user_id, worksheet_uuid)
-        _set_bundle_permissions([{
+        set_bundle_permissions([{
             'object_uuid': bundle_uuid,
             'group_uuid': p['group_uuid'],
             'permission': p['permission'],
@@ -271,7 +278,7 @@ def create_bundles():
 
 
 @patch('/bundles', apply=AuthenticatedPlugin())
-def update_bundles():
+def _update_bundles():
     """
     Bulk update bundles.
     """
@@ -299,7 +306,7 @@ def update_bundles():
 
 
 @delete('/bundles', apply=AuthenticatedPlugin())
-def delete_bundles():
+def _delete_bundles():
     """
     Delete the bundles specified.
     If |force|, allow deletion of bundles that have descendants or that appear across multiple worksheets.
@@ -312,28 +319,28 @@ def delete_bundles():
     recursive = query_get_bool('recursive', default=False)
     data_only = query_get_bool('data-only', default=False)
     dry_run = query_get_bool('dry-run', default=False)
-    deleted_uuids = _delete_bundles(uuids, force=force, recursive=recursive,
-                                    data_only=data_only, dry_run=dry_run)
+    deleted_uuids = delete_bundles(uuids, force=force, recursive=recursive,
+                                   data_only=data_only, dry_run=dry_run)
 
     # Return list of deleted ids as meta
     return json_api_meta({}, {'ids': deleted_uuids})
 
 
 @post('/bundle-permissions', apply=AuthenticatedPlugin())
-def set_bundle_permissions():
+def _set_bundle_permissions():
     """
     Bulk set bundle permissions.
     """
     new_permissions = BundlePermissionSchema(
         strict=True, many=True,
     ).load(request.json).data
-    _set_bundle_permissions(new_permissions)
+    set_bundle_permissions(new_permissions)
     return BundlePermissionSchema(many=True).dump(new_permissions).data
 
 
 @get('/bundles/<uuid:re:%s>/contents/blob/' % spec_util.UUID_STR)
 @get('/bundles/<uuid:re:%s>/contents/blob/<path:path>' % spec_util.UUID_STR)
-def fetch_bundle_contents_blob(uuid, path=''):
+def _fetch_bundle_contents_blob(uuid, path=''):
     """
     API to download the contents of a bundle or a subpath within a bundle.
 
@@ -387,7 +394,7 @@ def fetch_bundle_contents_blob(uuid, path=''):
 
 @put('/bundles/<uuid:re:%s>/contents/blob/' % spec_util.UUID_STR,
      apply=AuthenticatedPlugin())
-def update_bundle_contents_blob(uuid):
+def _update_bundle_contents_blob(uuid):
     """
     Update the contents of the given running or uploading bundle.
 
@@ -415,6 +422,11 @@ def update_bundle_contents_blob(uuid):
         raise
 
 
+#############################################################
+#  BUNDLE HELPER FUNCTIONS
+#############################################################
+
+
 def request_accepts_gzip_encoding():
     # See rules for parsing here: https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html
     # Browsers silently decode gzipped files, so we save some bandwidth.
@@ -428,3 +440,213 @@ def request_accepts_gzip_encoding():
             else:
                 return True
     return False
+
+
+def resolve_bundle_specs(worksheet_uuid, bundle_specs):
+    return [resolve_bundle_spec(worksheet_uuid, bundle_spec)
+            for bundle_spec in bundle_specs]
+
+
+def resolve_bundle_spec(worksheet_uuid, bundle_spec):
+    if '/' in bundle_spec:  # <worksheet_spec>/<bundle_spec>
+        # Shift to new worksheet
+        worksheet_spec, bundle_spec = bundle_spec.split('/', 1)
+        worksheet_uuid = get_worksheet_uuid(worksheet_uuid, worksheet_spec)
+
+    return canonicalize.get_bundle_uuid(local.model, request.user.user_id,
+                                        worksheet_uuid, bundle_spec)
+
+
+def mask_bundle(bundle_info):
+    """
+    Return a copy of the bundle_info dict that hides all fields except 'uuid'.
+    """
+    return {
+        'uuid': bundle_info['uuid'],
+        'bundle_type': PrivateBundle.BUNDLE_TYPE,
+        'owner_id': None,
+        'command': None,
+        'data_hash': None,
+        'state': None,
+        'metadata': [{
+            'metadata_key': 'name',
+            'metadata_value': '<private>',
+        }],
+        'dependencies': [],
+    }
+
+
+@local_bundle_client_compatible
+def delete_bundles(local, request, uuids, force, recursive, data_only, dry_run):
+    """
+    Delete the bundles specified by |uuids|.
+    If |force|, allow deletion of bundles that have descendants or that appear across multiple worksheets.
+    If |recursive|, add all bundles downstream too.
+    If |data_only|, only remove from the bundle store, not the bundle metadata.
+    """
+    relevant_uuids = local.model.get_self_and_descendants(uuids, depth=sys.maxint)
+    if not recursive:
+        # If any descendants exist, then we only delete uuids if force = True.
+        if (not force) and set(uuids) != set(relevant_uuids):
+            relevant = local.model.batch_get_bundles(uuid=(set(relevant_uuids) - set(uuids)))
+            raise UsageError('Can\'t delete bundles %s because the following bundles depend on them:\n  %s' % (
+                ' '.join(uuids),
+                '\n  '.join(bundle.simple_str() for bundle in relevant),
+            ))
+        relevant_uuids = uuids
+    check_bundles_have_all_permission(local.model, request.user, relevant_uuids)
+
+    # Make sure we don't delete bundles which are active.
+    states = local.model.get_bundle_states(uuids)
+    active_uuids = [uuid for (uuid, state) in states.items() if state in State.ACTIVE_STATES]
+    if len(active_uuids) > 0:
+        raise UsageError('Can\'t delete bundles: %s. ' % (' '.join(active_uuids)) +
+                         'For run bundles, kill them first. ' +
+                         'Bundles stuck not running will eventually ' +
+                         'automatically be moved to a state where they ' +
+                         'can be deleted.')
+
+    # Make sure that bundles are not referenced in multiple places (otherwise, it's very dangerous)
+    result = local.model.get_host_worksheet_uuids(relevant_uuids)
+    for uuid, host_worksheet_uuids in result.items():
+        worksheets = local.model.batch_get_worksheets(fetch_items=False, uuid=host_worksheet_uuids)
+        frozen_worksheets = [worksheet for worksheet in worksheets if worksheet.frozen]
+        if len(frozen_worksheets) > 0:
+            raise UsageError("Can't delete bundle %s because it appears in frozen worksheets "
+                             "(need to delete worksheet first):\n  %s" %
+                             (uuid, '\n  '.join(worksheet.simple_str() for worksheet in frozen_worksheets)))
+        if not force and len(host_worksheet_uuids) > 1:
+            raise UsageError("Can't delete bundle %s because it appears in multiple worksheets "
+                             "(--force to override):\n  %s" %
+                             (uuid, '\n  '.join(worksheet.simple_str() for worksheet in worksheets)))
+
+    # Delete the actual bundle
+    if not dry_run:
+        if data_only:
+            # Just remove references to the data hashes
+            local.model.remove_data_hash_references(relevant_uuids)
+        else:
+            # Actually delete the bundle
+            local.model.delete_bundles(relevant_uuids)
+
+        # Update user statistics
+        local.model.update_user_disk_used(request.user.user_id)
+
+    # Delete the data.
+    for uuid in relevant_uuids:
+        # check first is needs to be deleted
+        bundle_location = local.bundle_store.get_bundle_location(uuid)
+        if os.path.lexists(bundle_location):
+            local.bundle_store.cleanup(uuid, dry_run)
+
+    return relevant_uuids
+
+
+@local_bundle_client_compatible
+def get_bundle_infos(local, request, uuids, get_children=False,
+                     get_host_worksheets=False, get_permissions=False):
+    """
+    get_children, get_host_worksheets, get_permissions:
+        whether we want to return more detailed information.
+    Return map from bundle uuid to info.
+    """
+    if len(uuids) == 0:
+        return {}
+    bundles = local.model.batch_get_bundles(uuid=uuids)
+    bundle_dict = {bundle.uuid: bundle_util.bundle_to_bundle_info(local.model, bundle) for bundle in bundles}
+
+    # Filter out bundles that we don't have read permission on
+    def select_unreadable_bundles(uuids):
+        permissions = local.model.get_user_bundle_permissions(request.user.user_id, uuids, local.model.get_bundle_owner_ids(uuids))
+        return [uuid for uuid, permission in permissions.items() if permission < GROUP_OBJECT_PERMISSION_READ]
+
+    def select_unreadable_worksheets(uuids):
+        permissions = local.model.get_user_worksheet_permissions(request.user.user_id, uuids, local.model.get_worksheet_owner_ids(uuids))
+        return [uuid for uuid, permission in permissions.items() if permission < GROUP_OBJECT_PERMISSION_READ]
+
+    # Lookup the user names of all the owners
+    user_ids = [info['owner_id'] for info in bundle_dict.values()]
+    users = local.model.get_users(user_ids=user_ids) if len(user_ids) > 0 else []
+    users = {u.user_id: u for u in users}
+    if users:
+        for info in bundle_dict.values():
+            user = users[info['owner_id']]
+            info['owner_name'] = user.user_name if user else None
+            info['owner'] = '%s(%s)' % (info['owner_name'], info['owner_id'])
+
+    # Mask bundles that we can't access
+    for uuid in select_unreadable_bundles(uuids):
+        if uuid in bundle_dict:
+            bundle_dict[uuid] = mask_bundle(bundle_dict[uuid])
+
+    if get_children:
+        result = local.model.get_children_uuids(uuids)
+        # Gather all children bundle uuids
+        children_uuids = [uuid for l in result.values() for uuid in l]
+        unreadable = set(select_unreadable_bundles(children_uuids))
+        # Lookup bundle names
+        names = local.model.get_bundle_names(children_uuids)
+        # Fill in info
+        for uuid, info in bundle_dict.items():
+            info['children'] = [
+                {
+                    'uuid': child_uuid,
+                    'metadata': {'name': names[child_uuid]}
+                }
+                for child_uuid in result[uuid] if child_uuid not in unreadable]
+
+    if get_host_worksheets:
+        # bundle_uuids -> list of worksheet_uuids
+        result = local.model.get_host_worksheet_uuids(uuids)
+        # Gather all worksheet uuids
+        worksheet_uuids = [uuid for l in result.values() for uuid in l]
+        unreadable = set(select_unreadable_worksheets(worksheet_uuids))
+        worksheet_uuids = [uuid for uuid in worksheet_uuids if uuid not in unreadable]
+        # Lookup names
+        worksheets = dict(
+            (worksheet.uuid, worksheet)
+            for worksheet in local.model.batch_get_worksheets(
+                fetch_items=False,
+                uuid=worksheet_uuids))
+        # Fill the info
+        for uuid, info in bundle_dict.items():
+            info['host_worksheets'] = [
+                {
+                    'uuid': worksheet_uuid,
+                    'name': worksheets[worksheet_uuid].name
+                }
+                for worksheet_uuid in result[uuid]
+                if worksheet_uuid not in unreadable]
+
+    if get_permissions:
+        # Fill the info
+        group_result = local.model.batch_get_group_bundle_permissions(request.user.user_id, uuids)
+        result = local.model.get_user_bundle_permissions(request.user.user_id, uuids, local.model.get_bundle_owner_ids(uuids))
+        for uuid, info in bundle_dict.items():
+            info['group_permissions'] = group_result[uuid]
+            info['permission'] = result[uuid]
+
+    return bundle_dict
+
+
+@local_bundle_client_compatible
+def resolve_owner_in_keywords(local, request, keywords):
+    # Resolve references to owner ids
+    def resolve(keyword):
+        # Example: owner=codalab => owner_id=0
+        m = re.match('owner=(.+)', keyword)
+        if not m:
+            return keyword
+        return 'owner_id=%s' % getattr(local.model.get_user(username=m.group(1)), 'user_id', 'x')
+    return map(resolve, keywords)
+
+
+@local_bundle_client_compatible
+def set_bundle_permissions(local, request, new_permissions):
+    # Check if current user has permission to set bundle permissions
+    check_bundles_have_all_permission(
+        local.model, request.user, [p['object_uuid'] for p in new_permissions])
+    # Sequentially set bundle permissions
+    for p in new_permissions:
+        local.model.set_group_bundle_permission(
+            p['group_uuid'], p['object_uuid'], p['permission'])
